% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_config.R
\name{config_create}
\alias{config_create}
\alias{config_write}
\alias{config_validate}
\title{config_create}
\usage{
config_create(
  inDir = NULL,
  f1_suffix = NULL,
  f2_suffix = NULL,
  outDir = NULL,
  reference = NULL,
  starIndex = NULL,
  suppaRef = NULL,
  gtf = NULL,
  fasta = NULL,
  rDNA_starIndex = NULL,
  PE = NULL,
  SE = NULL,
  noSub = NULL,
  pipeline = NULL,
  jobDir = NULL,
  scriptLocation = NULL,
  docker = NULL,
  singularity = NULL
)

config_write(config_list, config_file, overwrite = FALSE)

config_validate(config_file, extra_args = NULL, work_dir = NULL)
}
\arguments{
\item{inDir}{Directory with all .fastq.gz files. Default is current directory.}

\item{f1_suffix}{Suffix of R1 files. Default is _R1_001.fastq.gz}

\item{f2_suffix}{Suffix of R2 files. Default is _R2_001.fastq.gz. Only required if data is PE.}

\item{outDir}{Output directory. Default is TAP_output in current directory.}

\item{reference}{Parent directory for all reference directories. The location with output from setup_new_reference.sh}

\item{starIndex}{Directory with STAR index files. Only required if \code{reference} is not provided or you wish to override.}

\item{suppaRef}{Directory with suppa2 index files. Only required if \code{reference} is not provided or you wish to override.}

\item{gtf}{File, gtf or gff, with gene annotation. Only required if \code{reference} is not provided or you wish to override.}

\item{fasta}{Organism genomic sequence fasta. Only required if \code{reference} is not provided or you wish to override.}

\item{rDNA_starIndex}{Directory with STAR index for rDNA. Only required if you wish to quantify rDNA read composition.}

\item{PE}{Set TRUE if data is PE. RNAseq will default to PE.}

\item{SE}{Set TRUE if data is SE. ChIPseq will default to SE.}

\item{noSub}{Set TRUE if bash should be used for all commands in serial instead of sbatch.}

\item{pipeline}{Pipeline script to use. Only required if submission script is not in the same directory as default of "rnaseq_pipeline.sh".}

\item{jobDir}{Directory with all job scripts. Only required if submission script is not in the same directory as default of "deployed_job_scripts".}

\item{docker}{If you want to use docker, docker image to use. i.e. jrboyd/tap after using "setup_scripts/pull_docker_image.sh"}

\item{singularity}{If you want to use singularity, singularity .sif file to use. i.e. tap_latest.sif after using "setup_scripts/pull_singularity_image.sh"}

\item{config_list}{list returned from \link{config_create}}

\item{config_file}{A configuration file for TAP.}

\item{overwrite}{If TRUE, existing config file will be overwritted. Default is FALSE.}

\item{scriptLocatio}{Deprecated. Don't use.}
}
\value{
list of config lines and fastq lines

path to configuration file.

A data.frame with the metadata elements parsed for each found fastq file.
}
\description{
Use with output from \link{config_create}. Modify fastq names as necessary.
}
\examples{
cfg = config_create(
  inDir = example_honeybee_input(),
  f1_suffix = "_1.fastq.gz", f2_suffix = "_2.fastq.gz",
  outDir = paste0(example_honeybee_output(), ".test2"),
  singularity = "~/lab_shared/scripts/TAP/tap_latest.sif",
  reference = example_honeybee_reference()
)
config_write(cfg, "test_config.csv")
config_validate("test_config.csv")

cfg_rename = config_create(
  inDir = example_honeybee_input(),
  f1_suffix = "_1.fastq.gz", f2_suffix = "_2.fastq.gz",
  outDir = paste0(example_honeybee_output(), ".rename2"),
  singularity = "~/lab_shared/scripts/TAP/tap_latest.sif",
  reference = example_honeybee_reference()
)
cfg_fq = cfg_rename$fastq_lines
colnames(cfg_fq) = c("file", "srr")
meta_df = example_honeybee_metadata()
cfg_fq = merge(meta_df, cfg_fq, by = "srr")
cfg_fq = cfg_fq[, c("file", "name")]
cfg_rename$fastq_lines = cfg_fq

config_write(cfg_rename, "rename_config.csv")
config_validate("rename_config.csv")
}
