% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_config.R
\name{config_validate}
\alias{config_validate}
\title{Validate config file for running TAP on RNAseq}
\usage{
config_validate(config_file, extra_args = NULL, wd = NULL)
}
\arguments{
\item{config_file}{A configuration file for TAP.}
}
\value{
A data.frame with the metadata elements parsed for each found fastq file.
}
\description{
Validate config file for running TAP on RNAseq
}
\examples{
cfgs = dir("~/lab_shared/scripts/TAP/testing/test_configs/", full.names = TRUE)
cfgs = cfgs[!grepl("chip", cfgs)]
names(cfgs) = basename(cfgs)
cfgs = as.list(cfgs)

test_dir = "/slipstream/home/joeboyd/lab_shared/scripts/TAP/testing"
cmd_extra = paste("-i", file.path(test_dir, "test_data/fastq_rnaseq_PE"), "-ref", file.path(test_dir, "references/dm6"))
config_validate(cfgs$test_dm6_config.basic.csv, extra_args = cmd_extra)

config_validate(cfgs$test_dm6_config.basic.csv, wd = "~/lab_shared/scripts/TAP/testing/test_data/fastq_rnaseq_PE/")

config_validate(cfgs$test_dm6_config.params.csv, wd = "~/lab_shared/scripts/TAP")

config_validate(cfgs$test_dm6_config.pool.csv, wd = "~/lab_shared/scripts/TAP/testing/test_data/fastq_rnaseq_PE/")

config_validate(cfgs$test_dm6_config.rDNA_only.csv, wd = "~/lab_shared/scripts/TAP")

config_validate(cfgs$test_dm6_config.rename.csv, wd = "~/lab_shared/scripts/TAP/testing/test_data/fastq_rnaseq_PE/")

config_validate(cfgs$test_dm6_config.SE.csv, wd = "~/lab_shared/scripts/TAP")

}
