% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scrape_encode.R
\name{ENCODE_download_files}
\alias{ENCODE_download_files}
\title{ENCODE_download_files}
\usage{
ENCODE_download_files(dt, output_location = getwd())
}
\arguments{
\item{dt}{Output of ENCODE_get_file_info with file_name attribute set. Inspect this data.table carefully and filter out unwanted files.}

\item{output_location}{Optional output directory. Will use current directory as default.}
}
\value{
The output location is invisibly returned.
}
\description{
ENCODE_download_files
}
\examples{
dt = ENCODE_get_file_info()
#just to avoid downloading anything too big
dt = dt[grepl("Mb", file_size)]
dt = dt[sample(seq_len(nrow(dt)), 5), ]
dt$file_format
dt[, file_name := paste0("test_", seq(.N), ".", file_format)]
dt[, .(file_name)]
ENCODE_download_files(dt)
}
