% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scrape_encode.R
\name{ENCODE_get_file_info}
\alias{ENCODE_get_file_info}
\title{ENCODE_get_file_info}
\usage{
ENCODE_get_file_info(
  request_organisms = NULL,
  request_cell_lines = NULL,
  request_target_chipseq = NULL,
  request_file_types = NULL
)
}
\arguments{
\item{request_organisms}{Organisms to select for.}

\item{request_cell_lines}{Cell lines to select for.}

\item{request_target_chipseq}{ChIPseq target to select for.}

\item{request_file_types}{File types to select for.}
}
\value{
A data.table output from \code{\link[ENCODExplorer:queryEncode]{ENCODExplorer::queryEncode()}}. Filter and then pass to \code{\link[=ENCODE_download_files]{ENCODE_download_files()}}
}
\description{
ENCODE_get_file_info
}
\examples{
dt = ENCODE_get_file_info()
dt
}
