% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_config}
\alias{example_honeybee_config}
\title{example_honeybee_config}
\usage{
example_honeybee_config()
}
\value{
Path to example TAP output
}
\description{
example_honeybee_config
}
\examples{
example_honeybee_config()
}
