% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_input}
\alias{example_honeybee_input}
\title{example_honeybee_input}
\usage{
example_honeybee_input()
}
\value{
Path to example TAP output
}
\description{
example_honeybee_input
}
\examples{
example_honeybee_input()
}
