% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_metadata}
\alias{example_honeybee_metadata}
\title{exampple_honeybee_metadata}
\usage{
example_honeybee_metadata()
}
\value{
data.frame with honeybee SRR metadata.
}
\description{
exampple_honeybee_metadata
}
\examples{
exampple_honeybee_metadata()
}
