% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_output.in_progress}
\alias{example_honeybee_output.in_progress}
\title{example_honeybee_output.in_progress}
\usage{
example_honeybee_output.in_progress()
}
\value{
Path to example TAP output
}
\description{
example_honeybee_output.in_progress
}
\examples{
example_honeybee_output.in_progress()
}
