% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_output}
\alias{example_honeybee_output}
\title{example_honeybee_output}
\usage{
example_honeybee_output()
}
\value{
Path to example TAP output
}
\description{
example_honeybee_output
}
\examples{
example_honeybee_output()
}
