% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{example_honeybee_reference}
\alias{example_honeybee_reference}
\title{example_honeybee_reference}
\usage{
example_honeybee_reference()
}
\value{
Path to example TAP reference
}
\description{
example_honeybee_reference
}
\examples{
example_honeybee_reference()
}
