% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scrape_encode.R
\name{fetch_encode_files}
\alias{fetch_encode_files}
\title{fetch_encode_files}
\usage{
fetch_encode_files(dt, out_dir = getwd())
}
\arguments{
\item{dt}{Output of get_encode_table with file_name attribute set.}

\item{out_dir}{Optional output directory. Will use current directory as default.}
}
\description{
fetch_encode_files
}
\examples{
dt = get_encode_table()
#just to avoid downloading anything too big
dt = dt[grepl("Mb", file_size)]
dt = dt[sample(seq_len(nrow(dt)), 5), ]
dt$file_format
dt[, file_name := paste0("test_", seq(.N), ".", file_format)]
dt[, .(file_name)]
fetch_encode_files(dt)
}
