% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuctions_ucsc_tracks_app.R
\name{launch_UCSC_tracks_app}
\alias{launch_UCSC_tracks_app}
\title{launch_UCSC_tracks_app}
\usage{
launch_UCSC_tracks_app(track_hosting_dir, base_host_path, base_url)
}
\arguments{
\item{track_hosting_dir}{The directory where track files are available via a
public URL. To conveniently place pipeline outputs in a public location,
use the \link{stage_output_for_UCSC_tracks} function.}

\item{base_host_path}{The portion of the \code{track_hosting_dir} to replace with
\code{base_url} to generate a valid URL.}

\item{base_url}{The public URL pointing to \code{base_host_path} locally.}
}
\value{
A shiny application is launched. Nothing is returned.
}
\description{
launch_UCSC_tracks_app
}
\examples{
work_dir = "/slipstream/home/joeboyd/public_files/honeybee"
track_hosting_dir = "/slipstream/home/joeboyd/public_files/honeybee"
base_host_path = "/slipstream/galaxy/production/galaxy-dist"
if(!grepl(base_host_path, track_hosting_dir)){
  sp = strsplit(track_hosting_dir, "/")[[1]]
  user = sp[which(sp == "home") + 1]
  track_hosting_dir = paste0("/slipstream/galaxy/production/galaxy-dist/static/UCSCtracks/", user, "_files")
  if(!dir.exists(track_hosting_dir)){
    "track_hosting_dir is not a sub-directory of base_host_path and could not convert to a UVM galaxy path."
  }
}
launch_UCSC_tracks_app(track_hosting_dir = track_hosting_dir,
                       base_host_path,
                       base_url = "https://galaxy.med.uvm.edu")
}
