% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuctions_ucsc_tracks_app.R
\name{launch_UCSC_tracks_app.UVM_galaxy}
\alias{launch_UCSC_tracks_app.UVM_galaxy}
\title{launch_UCSC_tracks_app.UVM_galaxy}
\usage{
launch_UCSC_tracks_app.UVM_galaxy(track_hosting_dir)
}
\arguments{
\item{track_hosting_dir}{The directory where track files are available via a public URL. To conveniently place pipeline outputs in a public location, use the \link{stage_output_for_UCSC_tracks} function.}
}
\value{
A shiny application is launched. Nothing is returned.
}
\description{
This variant of launch_UCSC_tracks_app has convenient presets for users of the Galaxy server at UVM and should not be used by others.
}
\examples{
launch_UCSC_tracks_app.UVM_galaxy("/slipstream/home/joeboyd/public_files/honeybee")
}
