% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{load_norm_counts}
\alias{load_norm_counts}
\title{xeno_rna.load_norm_counts}
\usage{
load_norm_counts(
  work_dir,
  library_type = NULL,
  name_composition = NULL,
  just_check_library_type = FALSE,
  gtf_file = NULL
)
}
\value{
wide matrix of RNA counts, RPM normalized after trimming .01 quantile
}
\description{
xeno_rna.load_norm_counts
}
\examples{
load_norm_counts()
}
