% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{load_transcript_reference}
\alias{load_transcript_reference}
\title{load_transcript_reference}
\usage{
load_transcript_reference(gtf_file)
}
\value{
A GRanges object with exonic information.
}
\description{
Loads transcript reference from the same .gtf file used to run TAP.
}
\details{
This loads all transcript entries. Use \link{load_gene_reference} if you only need basic gene information or \link{load_exon} reference for more detail.
}
\examples{
ref_dir = example_honeybee_reference()
gtf_file = file.path(ref_dir, "GTF/current.gtf")
load_transcript_reference(gtf_file)
# for simplicity you can just specify the base path to the reference used to run TAP.
load_transcript_reference(ref_dir)
}
