% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_csaw.R
\name{make_contrasts}
\alias{make_contrasts}
\title{make_contrasts}
\usage{
make_contrasts(config_df, groups_var = "cell", vary_var = "treatment")
}
\arguments{
\item{config_df}{Data.frame containing metadata with grouping variables for
differential analysis.}

\item{groups_var}{Defines column with grouping assignments within which
comparisons will be made.}

\item{vary_var}{Defines column that identifies treatment variable between
which comparisons will be made.}
}
\value{
A list with 2 named elements. A "contrast" matrix produced by limma::makeContrasts and the "design" matrix that assigns items in config to groups.
}
\description{
make_contrasts
}
\examples{
cfg_file = system.file(package = "TAPhelpR", "extdata/test_diff_config_cnr.csv", mustWork = TRUE)
cfg_df = read.table(cfg_file, sep = ",", header = TRUE)
cfg_df = subset(cfg_df, mark == "H3K27ac")
cfg_df
make_contrasts(cfg_df, groups_var = "background", vary_var = "treatment")
}
