% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output_report.R
\name{report_completion}
\alias{report_completion}
\alias{report_progress}
\alias{report_progress_plot}
\alias{report_errors}
\title{report_completion}
\usage{
report_completion(tap_out)

report_progress(tap_out)

report_progress_plot(tap_out)

report_errors(tap_out)
}
\arguments{
\item{tap_out}{TAP output directory, either in progress or completed}
}
\value{
data.frame reporting completion status per sample.

Report additional details for each sample.

Plot of job progress.

Locate jobs that experienced errors and return relevant log files.
}
\description{
report_completion

report_progress

report_progress_plot

report_errors
}
\examples{
tap_out = example_honeybee_output()
report_completion(tap_out)
report_progress(tap_out)
report_progress_plot(tap_out)
report_errors(tap_out)

tap_out = example_honeybee_output.in_progress()
report_completion(tap_out)
report_progress(tap_out)
report_progress_plot(tap_out)
report_errors(tap_out)
}
