% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scrape_geo.R
\name{scrape_geo}
\alias{scrape_geo}
\title{scrape_geo}
\usage{
scrape_geo(
  GSE_id,
  gsm_override = NULL,
  gsm_todo = NULL,
  gsm_skip = NULL,
  do_srr = TRUE,
  debug = FALSE,
  chara = c("cancer type", "cell type", "genotype", "passages", "antibody"),
  key_str = c("Title", rep("Characteristics", length(chara))),
  key_idx = append(list(5), chara)
)
}
\description{
Characteristics
  cell type: stromal adherent pre-B cell progenitors
  genotype: IKDN CD2 Cre
  passages: 3 to 8
  antibody: Abnova/H00001879-M01
}
\examples{
my_gse = "GSE86897"
gse_dt = scrape_geo(my_gse)
gse_dt
stopifnot(!any(duplicated(gse_dt$title)))
dump_script = system.file("extdata/fasterq_dump_wrapper.sh", package = "TAPhelpR", mustWork = TRUE)
srr_tofetch =
library(data.table)

gse_dt[, title := sub("Badh", "_Bad", title)]
gse_dt[, title := sub("InputControl_", "input_", title)]
gse_dt[, title := sub("pe_", "pre_", title)]
gse_dt[, title := sub("pre_", "pre", title)]
gse_dt[, title := sub("Pre", "pre", title)]
gse_dt[, title := sub("pre_", "pre", title)]
gse_dt[, title := sub("pre", "_pre", title)]
gse_dt[, title := gsub("_+", "_", title)]
gse_dt[, title := sub("H3K27Ac", "H3K27ac", title)]
gse_dt[, c("mark", "treatment", "cell", "method") := tstrsplit(title, "_", keep = 1:4)]

table(gse_dt$mark)
table(gse_dt$treatment)
table(gse_dt$cell)
gse_dt$cell = "preBad"
gse_dt[, treatment := sub("IkDN", "IKDN", treatment)]

table(gse_dt$mark)
table(gse_dt$treatment)
table(gse_dt$cell)

gse_dt[, prefix := paste("mouse", cell, mark, treatment, sep = "_")]

fastq_prefixes = gse_dt$prefix
out_dir = "zhang_IK_SE/fastqs"
fetch_cmds = submit_sra_fetch(gse_dt$srr, gse_dt$prefix, out_dir)#, docker = "jrboyd/tap")
writeLines(
sapply(fetch_cmds, system),
"submit_fetch_zhange.sh")
}
