% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{set_suppa_path}
\alias{set_suppa_path}
\title{set_suppa_path}
\usage{
set_suppa_path(suppa_path)
}
\arguments{
\item{suppa_path}{Path to suppa.py. Will be used for all suppa_* functions.}
}
\value{
Invisibly returns TRUE.
}
\description{
set_suppa_path
}
\examples{
set_suppa_path("/path/to/suppa.py")
}
