% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{setup_bam_files.transcriptome}
\alias{setup_bam_files.transcriptome}
\title{setup_bam_files.transcriptome}
\usage{
setup_bam_files.transcriptome(wd, var_map = NULL)
}
\arguments{
\item{var_map}{}
}
\description{
setup_bam_files.transcriptome
}
