% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_salmon_tpm.R
\name{setup_tpm_files}
\alias{setup_tpm_files}
\title{setup_tpm_files}
\usage{
setup_tpm_files(work_dir, variable_map = NULL)
}
\arguments{
\item{work_dir}{TAP output directory path}

\item{variable_map}{name of variables encoded in TAP output prefixes. Delimited by "_" and ".".}
}
\value{
data.table of files path to tpm files with metadata optionally extracted according to \code{variable_map}.
}
\description{
setup_tpm_files
}
\examples{
tap_out = "~/R_workspace.combined/TAPhelpR.data/honeybee_TAP_output"
tpm_files = setup_tpm_files(tap_out, variable_map = c("day", "sex", "rep"))
mat = load_matrix_from_tpm(tpm_files$file)
head(mat)
}
