% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\name{stage_output_for_UCSC_tracks}
\alias{stage_output_for_UCSC_tracks}
\title{stage_output_for_UCSC_tracks}
\usage{
stage_output_for_UCSC_tracks(pipeline_outputs_dir)
}
\arguments{
\item{pipeline_outputs_dir}{}
}
\description{
stage_output_for_UCSC_tracks
}
\examples{
config_file = "/slipstream/home/joeboyd/lab_shared/data/SF_SE_data"
config_file = "/slipstream/home/joeboyd/R_workspace.combined/chiptsne_projects/zhang_IK_SE/config_zhang_IK_SE.csv"
pipeline_outputs_dir = "/slipstream/home/joeboyd/lab_shared/data/SF_SE_data"
pipeline_outputs_dir = "/slipstream/home/joeboyd/R_workspace.combined/chiptsne_projects/zhang_IK_SE/output"


c("fe", "nu", "ns", "ru", "rs")

stage_output_for_UCSC_tracks("/slipstream/home/joeboyd/R_workspace.combined/chiptsne_projects/zhang_IK_SE/output")
stage_output_for_UCSC_tracks("/slipstream/home/joeboyd/lab_shared/data/SF_SE_data")
}
