% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diffsplice.R
\name{suppa_clusterEvents}
\alias{suppa_clusterEvents}
\title{suppa_clusterEvents}
\usage{
suppa_clusterEvents(
  diffSplice_output_location = "suppa2_output",
  PSI_todo = unlist(SPLICE_EVENTS),
  output_location = diffSplice_output_location
)
}
\arguments{
\item{diffSplice_output_location}{Working directory with output from suppa_joinFiles. Will be reused as output_location if not set.}

\item{PSI_todo}{Character vector of alternate events or isoform recognized by suppa2.  See variable SPLICE_EVENTS for valid values. Default will run all event types.}

\item{output_location}{Location to write output files to. Same as diffSplice_output_location if not specified.}
}
\value{
The output location is invisibly returned.
}
\description{
suppa_clusterEvents
}
\examples{
diffSplice_output_location = "suppa2_output"
suppa_clusterEvents(diffSplice_output_location)
}
