% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diffsplice.R
\name{suppa_diffSplice}
\alias{suppa_diffSplice}
\title{suppa_diffSplice}
\usage{
suppa_diffSplice(
  ref_location,
  joinFiles_output_location = "suppa2_output",
  PSI_todo = unlist(SPLICE_EVENTS),
  diffSplice_output_location = joinFiles_output_location
)
}
\arguments{
\item{ref_location}{Location of SUPPA2 directory in TAP reference.}

\item{joinFiles_output_location}{Working directory with output from \link{suppa_joinFiles}. Will be
reused as diffSplice_output_location if not set.}

\item{PSI_todo}{Character vector of alternate events or isoform recognized by
suppa2.  See variable SPLICE_EVENTS for valid values. Default will run all
event types.}

\item{diffSplice_output_location}{Location to write output files to. Same as joinFiles_output_location if not
specified.}
}
\value{
The output location is invisibly returned.
}
\description{
Run after \link{suppa_joinFiles} to compare all groups. If this would result in an
undesirable number of uninteresting pairwise comparisons, consider using
\link{suppa_diffSplice.within_group} insead.
}
\examples{
joinFiles_output_location = "suppa2_output"
ref_location = "~/indexes/HG38/SUPPA2"
suppa_diffSplice(joinFiles_output_location, ref_location)
}
