% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diffsplice.R
\name{suppa_diffSplice.within_group}
\alias{suppa_diffSplice.within_group}
\title{suppa_diffSplice.within_group}
\usage{
suppa_diffSplice.within_group(
  input_files,
  within_group,
  between_group,
  ref_location,
  joinFiles_output_location = "suppa2_output",
  PSI_todo = unlist(SPLICE_EVENTS),
  diffSplice_output_location = joinFiles_output_location
)
}
\arguments{
\item{input_files}{Paths to either .bam files or .salmon_quant directories in
a TAP output directory. See details for alternative formats and interaction
with other parameters.}

\item{ref_location}{Location of SUPPA2 directory in TAP reference.}

\item{joinFiles_output_location}{Working directory with output from \link{suppa_joinFiles}. Will be
reused as diffSplice_output_location if not set.}

\item{PSI_todo}{Character vector of alternate events or isoform recognized by
suppa2.  See variable SPLICE_EVENTS for valid values. Default will run all
event types.}

\item{diffSplice_output_location}{Location to write output files to. Same as joinFiles_output_location if not
specified.}

\item{by}{Only used when input_files is a data.frame. Must be a character
vector of attributes in input_files. Values of these attributes are used to
split input_files into groups to join.}
}
\value{
The output location is invisibly returned.
}
\description{
suppa_diffSplice.within_group
}
\examples{
tap_out = "~/R_workspace.combined/TAPhelpR.data/honeybee_TAP_output"
bam_files = setup_bam_files(tap_out, var_map = c("day", "sex", "rep"))

suppa_diffSplice.within_group(
  input_files = bam_files,
  within_group = "day",
  between_group = "sex",
  ref_location = "~/lab_shared/indexes/honeybee",
  PSI_todo = TAP_SPLICE_EVENTS$SkippingExon
)
}
