% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diffsplice.R
\name{suppa_joinFiles}
\alias{suppa_joinFiles}
\title{suppa_joinFiles}
\usage{
suppa_joinFiles(
  input_files,
  by = NULL,
  output_name = NULL,
  run_TPM = TRUE,
  PSI_todo = unlist(SPLICE_EVENTS),
  bam_suffix = ".Aligned.+",
  joinFiles_output_location = "suppa2_output"
)
}
\arguments{
\item{input_files}{Paths to either .bam files or .salmon_quant directories in
a TAP output directory. See details for alternative formats and interaction
with other parameters.}

\item{by}{Only used when input_files is a data.frame. Must be a character
vector of attributes in input_files. Values of these attributes are used to
split input_files into groups to join.}

\item{output_name}{File prefix for joinFiles outputs.}

\item{run_TPM}{If TRUE a joined tpm file will be generated. Default of TRUE.}

\item{PSI_todo}{Alternate splice event codes recognized by suppa2. See
TAP_SPLICE_EVENTS global variable for possible values.}

\item{bam_suffix}{Suffix that will be replaced in input_files to yield
.salmon_quant directories.}

\item{output_location}{Output location for joinFiles outputs. Will be created
if it does not exist. Default is "suppa2_output".}
}
\value{
The output location is invisibly returned.
}
\description{
This must be run prior to \link{suppa_diffSplice} to group data by comparison
group.
}
\details{
\code{input_files} can be one of 3 formats.
\enumerate{
\item A simple character vector of file paths. output_name is required in this case.
You'll need to call \code{suppa_joinFiles} multiple times in this scenario, once for
each comparison group.
\item A named list of character vectors of file paths. List names will be used for
output_name of each entry. If your list contains all groups then a single
suppa_joinFiles call is sufficient.
\item A data.frame where file paths are in "file" attribute or the first column.
The parameter \code{by} is required in this case but otherwise behaves similarly to
providing a list.
}
}
\examples{
wd = "/slipstream/home/dbgap/data/alignment_RNA-Seq/"
bam_files = find_bam_files(wd)
input_files = bam_files[1:5]
suppa_joinFiles(input_files, "test")

suppa_joinFiles(bam_files[1:5], "test1")
suppa_joinFiles(bam_files[6:10], "test2")
}
