% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{locate_files}
\alias{locate_files}
\alias{setup_bam_files}
\alias{setup_peak_files}
\alias{setup_count_files}
\alias{load_counts}
\title{locate_files}
\usage{
locate_files(wd, pattern)

setup_bam_files(wd, var_map = NULL)

setup_peak_files(wd, var_map = NULL)

setup_count_files(wd, var_map = NULL)

load_counts(
  wd,
  name_composition = NULL,
  just_check_library_type = FALSE,
  use_gene_name = TRUE,
  gtf_file = NULL
)
}
\arguments{
\item{pattern}{}
}
\value{
data.table/data.frame

data.table/data.frame

wide matrix of RNA counts
}
\description{
Creates data.table/data.frame with file paths and associated metadata.
}
\examples{
wd = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
locate_files(wd, "final.out$")
wd = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
setup_bam_files(wd)
#usage of var_map
setup_bam_files(wd, var_map = c("cell" = 1, "temp" = 2, "rep" = 3))
wd = "/slipstream_old/home/joeboyd/R_workspace/"
setup_peak_files(wd)
setup_peak_files(wd, var_map = c("cell" = 1, "mark" = 2, "rep" = 3))
wd = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
setup_count_files(wd)

setup_count_files(wd, var_map = c("cell" = 1, "temp" = 2, "rep" = 3))
wd = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
load_counts(wd)
load_counts(wd, name_composition = c(1, 2, 3, 4))
}
