% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{setup_files}
\alias{setup_files}
\alias{setup_bam_files}
\alias{setup_bam_files.transcriptome}
\alias{setup_peak_files}
\alias{setup_count_files}
\title{setup_files}
\usage{
setup_files(wd, pattern, var_map = NULL)

setup_bam_files(wd, var_map = NULL)

setup_bam_files.transcriptome(wd, var_map = NULL)

setup_peak_files(wd, var_map = NULL)

setup_count_files(wd, var_map = NULL)
}
\arguments{
\item{wd}{Path to TAP output directory}

\item{pattern}{File pattern to match (.bam, .samon_quant, .ReadsPerGene.out.tab, etc)}

\item{var_map}{Variables that can be extracted from file names. File names are assumed to be delimitted by "_" and/or ".". See details.}
}
\value{
data.table/data.frame

\code{var_map} encodes how metadata variables can be extracted from file names returned by matching the \code{pattern} in \code{wd}.

Let's use 6d_queen_A.Aligned.sortedByCoord.out.bam as an example file.

\code{var_map = c("day", "sex", "rep")} would extract 6d, queen, and A into columns "day", "sex", and "rep" respectively.

If you need to be more specific you could do:

\code{var_map = c("sex" = 2)} which would extract only the second item into column "sex".

Using this specification type we can get the same result as the first exmaple with:

\code{var_map = c("day" = 1, "sex" = 2, "rep" = 3)}
}
\description{
Creates data.table/data.frame with file paths and associated metadata.
}
\examples{
wd = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
setup_bam_files(wd)
#usage of var_map
setup_bam_files(wd, var_map = c("cell" = 1, "temp" = 2, "rep" = 3))
# general form
setup_bam_files(wd, var_map = c("cell" = 1, "temp" = 2, "rep" = 3), pattern = 'Aligned.sortedByCoord.out.bam$')
}
