% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_matrix_from_ReadsPerGene.out.tab}
\alias{load_matrix_from_ReadsPerGene.out.tab}
\title{load_matrix_from_ReadsPerGene.out.tab}
\usage{
load_matrix_from_ReadsPerGene.out.tab(files, lib_type = "guess")
}
\arguments{
\item{lib_type}{}
}
\description{
load_matrix_from_ReadsPerGene.out.tab
}
