% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output_report.R
\name{report_completion}
\alias{report_completion}
\alias{report_progress}
\alias{report_errors}
\title{report_completion}
\usage{
report_completion(tap_out)

report_progress(tap_out)

report_errors(tap_out)
}
\arguments{
\item{tap_out}{}
}
\description{
report_completion

report_progress

report_errors
}
\examples{
tap_out = "~/R_workspace.combined/TAPhelpR.data/honeybee_TAP_output.rename"
report_completion(tap_out)
report_progress(tap_out)
report_errors(tap_out)
}
