% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuctions_ucsc_tracks_app.R
\name{stage_output_for_UCSC_tracks}
\alias{stage_output_for_UCSC_tracks}
\title{stage_output_for_UCSC_tracks}
\usage{
stage_output_for_UCSC_tracks(
  pipeline_outputs_dir,
  track_hosting_dir,
  overwrite = FALSE
)
}
\arguments{
\item{pipeline_outputs_dir}{The directory where completed TAP outputs are
located.}

\item{track_hosting_dir}{The directory where track files are available via a
public URL. To conveniently place pipeline outputs in a public location,
use the \link{stage_output_for_UCSC_tracks} function.}

\item{overwrite}{Should files already present in \code{track_hosting_dir} be overwritten? Default is FALSE.}
}
\value{
\code{track_hosting_dir} is invisibly returned.
}
\description{
Uses symbolic links to make bigwig files available at a publicly hosted
location for later configuration for viewing on UCSC. Uses
\link{launch_UCSC_tracks_app} with the same \code{track_hosting_dir} to configure a
UCSC browser session.
}
\examples{
stage_output_for_UCSC_tracks("~/R_workspace.combined/TAPhelpR.data/honeybee_TAP_output", "~/public_files/honeybee_tracks")
}
