% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{load_counts}
\alias{load_counts}
\title{Creates data.table/data.frame with file paths and associated metadata.}
\usage{
load_counts(
  work_dir,
  library_type = NULL,
  name_composition = NULL,
  just_check_library_type = FALSE,
  gtf_file = NULL,
  name_attribute = "gene_name"
)
}
\arguments{
\item{work_dir}{Path to TAP output directory}

\item{library_type}{Strandedness of library. Should be one of "guess", "unstranded", "first", or "second". Default is "guess" and will result in attempting to automatically detect and apply library type.}

\item{name_composition}{Used to extract column names for final count matrix from basename of count files. Should be a vector of numbers defining positions in _ and/or . delimited basename. For instance, name_composition = c(1, 3) on a basename of donor1_august_liver.siteA would yield a column name of donor1_liver.}

\item{just_check_library_type}{If TRUE, counts will not be loaded into a matrix and instead a vector of library type guesses will be returned. For a more detailed look at individual libraries use \link{guess_lib_from_file} with show_plots = TRUE.}

\item{gtf_file}{Supply to convert gene_ids to gene_names (HGNC gene symbols). Either a path to a gtf file or GRanges object loaded from one with \link{load_gene_reference} or \link[rtracklayer:GFFFile-class]{rtracklayer::import.gff}.  The "gene_id" attribute will be matched to "gene_name" and aggregated by sum if required.}

\item{name_attribute}{The attribute in the gtf file to which gene_ids should be aggregated. Default of "gene_name" mostly works but some gtfs use "gene".}
}
\value{
wide matrix of RNA counts
}
\description{
Creates data.table/data.frame with file paths and associated metadata.
}
\examples{
work_dir = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
load_counts(work_dir)
load_counts(work_dir, name_composition = c(1, 2, 3, 4))
}
