% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_load.R
\name{load_counts}
\alias{load_counts}
\title{Creates data.table/data.frame with file paths and associated metadata.}
\usage{
load_counts(
  work_dir,
  lib_type = NULL,
  name_composition = NULL,
  just_check_library_type = FALSE,
  gtf_file = NULL
)
}
\arguments{
\item{gtf_file}{}
}
\value{
wide matrix of RNA counts
}
\description{
Creates data.table/data.frame with file paths and associated metadata.
}
\examples{
work_dir = "/slipstream_old/home/joeboyd/R_workspace/SFtapfly.data"
load_counts(work_dir)
load_counts(work_dir, name_composition = c(1, 2, 3, 4))
}
